refactor(core): extract dais-shared crate (signatures, AP types, E2EE) (#80)
Per CLIENT_REDESIGN.md §2/§3, create a `dais-shared` crate as the single
audited home for signing and crypto, shared by dais-core (Cloudflare Workers,
wasm32) and the new native client.

- dais-shared/signatures.rs: pure HTTP-Signature primitives, moved out of core.
- dais-shared/types.rs: ActivityPub wire types, moved out of core.
- dais-shared/e2ee.rs: NEW Rust port of cli/dais_cli/e2ee.py, kept wire-compatible
  (v:1 / AES-256-GCM / RSA-OAEP-256) so Python- and Rust-encrypted messages
  interoperate. E2EE previously lived only in the Python CLI, never in core.
- core's activitypub::{types,signatures} are now thin `pub use dais_shared::…`
  shims; core keeps only the HTTP-coupled fetch_actor_public_key.

dais-shared stays wasm-portable (no tokio/reqwest/sqlite). Verified: dais-shared
native+wasm green + 5 tests; core native+wasm green + 18 lib tests; actor worker
wasm build green (server build chain intact).
